Partially roll back commit f1f0f4523d5d084779a18d481bfe3a615e4ff1f9
The commit previously allowed null default values to be specified as 'null' (without quotes) to be marked as nullable, with a quoted null default value to be a string value of null. It was reported by several people on Slack as well as https://github.com/rainlab/builder-plugin/issues/295 that this did not work. This rolls back the changes for null, while keeping the changes for the boolean default values.